Updated meshoptimizer.

+++ 3rdparty/meshoptimizer/tools/basistoktx.cpp
@@ -172,8 +172,8 @@ std::string basisToKtx(const std::string& basis, bool srgb)
 	// supercompression global data
 	ktxBasisGlobalHeader sgd_header = {};
 	sgd_header.globalFlags = basis_header.m_flags;
-	sgd_header.endpointCount = basis_header.m_total_endpoints;
-	sgd_header.selectorCount = basis_header.m_total_selectors;
+	sgd_header.endpointCount = uint16_t(basis_header.m_total_endpoints);
+	sgd_header.selectorCount = uint16_t(basis_header.m_total_selectors);
 	sgd_header.endpointsByteLength = basis_header.m_endpoint_cb_file_size;
 	sgd_header.selectorsByteLength = basis_header.m_selector_cb_file_size;
 	sgd_header.tablesByteLength = basis_header.m_tables_file_size;
